Put html related filters into separate file
