+++ IPython/nbconvert/exporters/exporter.py
@@ -57,13 +57,14 @@ default_filters = {
         'ipython2python': filters.ipython2python,
         'posix_path': filters.posix_path,
         'markdown2latex': filters.markdown2latex,
-        'extended_markdown2latex': filters.markdown.extended_markdown2latex,
+        'extended_markdown2latex': filters.extended_markdown2latex,
         'markdown2rst': filters.markdown2rst,
         'comment_lines': filters.comment_lines,
         'strip_ansi': filters.strip_ansi,
         'strip_dollars': filters.strip_dollars,
         'strip_files_prefix': filters.strip_files_prefix,
         'html2text' : filters.html2text,
+        'html2latex' : filters.html2latex,
         'add_anchor': filters.add_anchor,
         'ansi2latex': filters.ansi2latex,
         'strip_math_space': filters.strip_math_space,
