+++ IPython/nbconvert/filters/__init__.py
@@ -4,3 +4,4 @@ from .highlight import *
 from .latex import *
 from .markdown import *
 from .strings import *
+from .html import *
