+++ IPython/nbconvert/filters/markdown.py
@@ -20,6 +20,7 @@ import sys
 import subprocess
 from HTMLParser import HTMLParser
 
+from .html import html2latex
 from IPython.nbconvert.utils.pandoc import pandoc
 
 #-----------------------------------------------------------------------------
@@ -29,7 +30,6 @@ from IPython.nbconvert.utils.pandoc import pandoc
 __all__ = [
     'markdown2html',
     'markdown2latex',
-    'html2latex',
     'markdown2rst',
     'extended_markdown2latex'
 ]
@@ -51,24 +51,6 @@ def markdown2latex(source):
       Output as returned by pandoc.
     """
     return pandoc(source, 'markdown', 'latex')
-
-def html2latex(source):
-    """Convert a html tag to LaTeX via pandoc.
-
-    This function will raise an error if pandoc is not installed.
-    Any error messages generated by pandoc are printed to stderr.
-
-    Parameters
-    ----------
-    source : string
-      Input string, assumed to be valid markdown.
-
-    Returns
-    -------
-    out : string
-      Output as returned by pandoc.
-    """
-    return pandoc(source, 'html', 'latex')
 
 def markdown2html(source):
     """Convert a markdown string to HTML via pandoc"""
